# BUG: MultiIndex intersection with sort=False  does not preserve order