--- a/asv_bench/benchmarks/multiindex_object.py
+++ b/asv_bench/benchmarks/multiindex_object.py
@@ -164,25 +164,38 @@ class SetOperations:
 
     params = [
         ("monotonic", "non_monotonic"),
+        ("datetime", "int", "string"),
         ("intersection", "union", "symmetric_difference"),
     ]
-    param_names = ["index_structure", "method"]
+    param_names = ["index_structure", "dtype", "method"]
 
-    def setup(self, index_structure, method):
+    def setup(self, index_structure, dtype, method):
         N = 10 ** 5
-        values = np.arange(N)
+        level1 = range(1000)
 
-        if index_structure == "monotonic":
-            mi = MultiIndex.from_arrays([values, values])
-        else:
-            mi = MultiIndex.from_arrays(
-                [np.concatenate([values[::2], values[1::2]]), values]
-            )
+        level2 = date_range(start="1/1/2000", periods=N // 1000)
+        dates_left = MultiIndex.from_product([level1, level2])
 
-        self.left = mi
-        self.right = mi[:-1]
+        level2 = range(N // 1000)
+        int_left = MultiIndex.from_product([level1, level2])
 
-    def time_operation(self, index_structure, method):
+        level2 = tm.makeStringIndex(N // 1000).values
+        str_left = MultiIndex.from_product([level1, level2])
+
+        data = {
+            "datetime": dates_left,
+            "int": int_left,
+            "string": str_left,
+        }
+
+        if index_structure == "non_monotonic":
+            data = {k: mi[::-1] for k, mi in data.items()}
+
+        data = {k: {"left": mi, "right": mi[:-1]} for k, mi in data.items()}
+        self.left = data[dtype]["left"]
+        self.right = data[dtype]["right"]
+
+    def time_operation(self, index_structure, dtype, method):
         getattr(self.left, method)(self.right)