--- a/doc/source/whatsnew/v1.1.0.rst
+++ b/doc/source/whatsnew/v1.1.0.rst
@@ -176,13 +176,13 @@ MultiIndex
         # Rows are now ordered as the requested keys
         df.loc[(['b', 'a'], [2, 1]), :]
 
-- Bug in :meth:`MultiIndex.intersection` was not preserving order when ``sort=False``. (:issue:`31325`)
+- Bug in :meth:`MultiIndex.intersection` was not guaranteed to preserve order when ``sort=False``. (:issue:`31325`)
 
 .. ipython:: python
 
         left = pd.MultiIndex.from_arrays([["b", "a"], [2, 1]])
         right = pd.MultiIndex.from_arrays([["a", "b", "c"], [1, 2, 3]])
-        # Common elements are now ordered by the left side
+        # Common elements are now guaranteed to be ordered by the left side
         left.intersection(right, sort=False)
 
 -