--- a/pandas/tests/indexes/multi/test_setops.py
+++ b/pandas/tests/indexes/multi/test_setops.py
@@ -19,18 +19,20 @@ def test_set_ops_error_cases(idx, case, sort, method):
 
 
 @pytest.mark.parametrize("sort", [None, False])
-def test_intersection_base(idx, sort):
+@pytest.mark.parametrize("klass", [MultiIndex, np.array, Series, list])
+def test_intersection_base(idx, sort, klass):
     first = idx[2::-1]  # first 3 elements reversed
     second = idx[:5]
 
-    array_like_cases = [klass(second.values) for klass in [np.array, Series, list]]
-    for case in [second, *array_like_cases]:
-        intersect = first.intersection(case, sort=sort)
-        if sort is None:
-            expected = first.sort_values()
-        else:
-            expected = first
-        tm.assert_index_equal(intersect, expected)
+    if klass is not MultiIndex:
+        second = klass(second.values)
+
+    intersect = first.intersection(second, sort=sort)
+    if sort is None:
+        expected = first.sort_values()
+    else:
+        expected = first
+    tm.assert_index_equal(intersect, expected)
 
     msg = "other must be a MultiIndex or a list of tuples"
     with pytest.raises(TypeError, match=msg):
@@ -38,18 +40,20 @@ def test_intersection_base(idx, sort):
 
 
 @pytest.mark.parametrize("sort", [None, False])
-def test_union_base(idx, sort):
+@pytest.mark.parametrize("klass", [MultiIndex, np.array, Series, list])
+def test_union_base(idx, sort, klass):
     first = idx[::-1]
     second = idx[:5]
 
-    array_like_cases = [klass(second.values) for klass in [np.array, Series, list]]
-    for case in [second, *array_like_cases]:
-        union = first.union(case, sort=sort)
-        if sort is None:
-            expected = first.sort_values()
-        else:
-            expected = first
-        tm.assert_index_equal(union, expected)
+    if klass is not MultiIndex:
+        second = klass(second.values)
+
+    union = first.union(second, sort=sort)
+    if sort is None:
+        expected = first.sort_values()
+    else:
+        expected = first
+    tm.assert_index_equal(union, expected)
 
     msg = "other must be a MultiIndex or a list of tuples"
     with pytest.raises(TypeError, match=msg):